+++ Zest/Controllers/CommunityModeratorsController.cs
@@ -14,11 +14,13 @@ namespace Zest.Controllers
     public class CommunityModeratorsController : Controller
     {
 		private readonly ICommunityModeratorService _communityModeratorService;
-	
-		public CommunityModeratorsController(ICommunityModeratorService communityModeratorService)
+	private readonly ICommunityService _communityService;
+		private readonly IAccountService _accountService;
+		public CommunityModeratorsController(ICommunityModeratorService communityModeratorService, ICommunityService communityService, IAccountService accountService)
 		{
 			this._communityModeratorService = communityModeratorService;
-			
+			_communityService=communityService;
+			_accountService=accountService;
 		}
 		[Authorize]
 		[Route("isModerator/{accountId}/{communityId}")]
@@ -33,6 +35,11 @@ namespace Zest.Controllers
 		[HttpPost]
 		public async Task<IActionResult> Add(string accountId, int communityId)
 		{
+			var isCandidate = await _communityModeratorService.IsModeratorCandidateAsync(accountId, communityId);
+			if (isCandidate)
+			{
+				return BadRequest("User is already candidate");
+			}
 			await _communityModeratorService.AddModeratorAsync(accountId, communityId);
 			return Ok();
 		}
@@ -40,16 +47,26 @@ namespace Zest.Controllers
 		[Authorize]
 		[Route("getModerators/{communityId}")]
 		[HttpGet]
-		public async Task<UserViewModel[]> GetModeratorsByCommunity(int communityId)
+		public async Task<ActionResult<UserViewModel[]>> GetModeratorsByCommunity(int communityId)
 		{
+			var doesCommunityExists = await _communityService.DoesExistAsync(communityId);
+			if (!doesCommunityExists)
+			{
+				return BadRequest("Community does not exists");
+			}
 			return await _communityModeratorService.GetModeratorsByCommunityAsync(communityId);
 		}
 
 		[Authorize]
 		[Route("getCandidates/{communityId}")]
 		[HttpGet]
-		public async Task<UserViewModel[]> GetModeratorCandidatesByCommunity(int communityId)
+		public async Task<ActionResult<UserViewModel[]>> GetModeratorCandidatesByCommunity(int communityId)
 		{
+			var doesCommunityExists = await _communityService.DoesExistAsync(communityId);
+			if (!doesCommunityExists)
+			{
+				return BadRequest("Community does not exists");
+			}
 			return await _communityModeratorService.GetModeratorCandidatesByCommunityAsync(communityId);
 		}
 
@@ -58,6 +75,16 @@ namespace Zest.Controllers
 		[HttpPost]
 		public async Task<IActionResult> ApproveCandidate(string accountId, int communityId)
 		{
+			var doesAccountExists = await _accountService.DoesExistAsync(accountId);
+			if (!doesAccountExists)
+			{
+				return BadRequest("Account does not exists");
+			}
+			var doesCommunityExists = await _communityService.DoesExistAsync(communityId);
+			if (!doesCommunityExists)
+			{
+				return BadRequest("Community does not exists");
+			}
 			await _communityModeratorService.ApproveCandidateAsync(accountId, communityId);
 			return Ok();
 		}
@@ -67,6 +94,16 @@ namespace Zest.Controllers
 		[HttpPost]
 		public async Task<IActionResult> RemoveModerator(string accountId, int communityId)
 		{
+			var doesAccountExists = await _accountService.DoesExistAsync(accountId);
+			if (!doesAccountExists)
+			{
+				return BadRequest("Account does not exists");
+			}
+			var doesCommunityExists = await _communityService.DoesExistAsync(communityId);
+			if (!doesCommunityExists)
+			{
+				return BadRequest("Community does not exists");
+			}
 			await _communityModeratorService.RemoveModeratorAsync(accountId, communityId);
 			return Ok();
 		}
